+++ .gitignore
@@ -132,3 +132,4 @@ dist
 ## database
 postgres-data/
 .DS_Store
+__pycache__/
