+++ api/Dockerfile
@@ -1,14 +1,20 @@
+# Use an official Python runtime as a parent image
 FROM python:3.9
 
+# Set the working directory in the container
 WORKDIR /app
 
-COPY requirements.txt .
+# Copy the current directory contents into the container at /app
+COPY . /app
 
+# Install any needed packages specified in requirements.txt
 RUN pip install --no-cache-dir -r requirements.txt
 
-# Adiciona o diretório pai ao PYTHONPATH
-ENV PYTHONPATH "${PYTHONPATH}:/app"
+# Make port 5100 available to the world outside this container
+EXPOSE 5100
 
-COPY . .
+# Define environment variable
+ENV PYTHONPATH /app
 
-CMD [ "python", "app.py" ]
+# Run app.py when the container launches
+CMD ["python", "app.py"]
