+++ api/app.py
@@ -1,13 +1,24 @@
 from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
+# Import SQLAlchemy instance
+from database import db
 from routes import routes_bp
-from models import db
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://postgres:password@taskevalpro-db/taskevalpro'
+
+# Link the db instance to the current app
 db.init_app(app)
 
-# Registrar blueprints com injeção de dependência
+# Since models are now imported after db and app have been defined,
+# it's safe to import them. Ensure that models are actually used or
+# referenced to avoid unused import warnings.
+with app.app_context():
+    from models.usuario import Usuario
+    from models.tarefa import Tarefa
+    from models.avaliacao import Avaliacao
+    db.create_all()  # Optional: Use only if you want to create tables at startup
+
+# Register blueprints for dependency injection
 app.register_blueprint(routes_bp, url_prefix='/api')
 
 if __name__ == '__main__':
