+++ api/models/__init__.py
@@ -1,0 +1,4 @@
+# api/models/__init__.py
+from .usuario import Usuario
+from .tarefa import Tarefa
+from .avaliacao import Avaliacao
