+++ api/models/usuario.py
@@ -1,0 +1,10 @@
+# api/models/usuario.py
+from database import db
+
+class Usuario(db.Model):
+    __tablename__ = 'usuarios'
+    id = db.Column(db.Integer, primary_key=True)
+    nome_usuario = db.Column(db.String(100), nullable=False)
+    email = db.Column(db.String(100), unique=True, nullable=False)
+    senha = db.Column(db.String(100), nullable=False)
+    papel = db.Column(db.String(50), nullable=False)
